chore: bump to muta v0.2.0-dev.0 (#84)

+++ services/governance/src/lib.rs
@@ -319,7 +319,7 @@ impl<SDK: ServiceSDK> GovernanceService<SDK> {
     }
 
     #[cfg(test)]
-    fn profits_len(&self) -> u32 {
+    fn profits_len(&self) -> u64 {
         self.profits.len()
     }
